+++ docs/internal/investigations/runtime-governor-poc/README.md
@@ -1,101 +1,58 @@
-# Runtime Governor - Proof of concept
+# Runtime Governor - Proof of Concept
 
-This directory contains the source code and Helm chart for Kyma Control Plane Runtime Governor.
+This document describes the scenario that will be developed to prove that [Dapr](https://dapr.io/) sidecars can be configured from a central place.
 
-## Overview
+## Reasons
 
-Runtime Governor exposes a simple REST API that returns Runtime configurations.
-The application periodically reloads configuration from the `runtime-governor-config` ConfigMap in the `kcp-poc` Namespace. The configuration file has the following form:
+Kyma Runtimes must consume fewer resources. It can be achieved by delegating some of Kyma's responsibilities
+to Dapr. Dapr injects sidecars into selected Pods where they fulfill their given tasks, such as state management or pub-sub.
+Configuration for these sidecars can be held in one central place, from where Kyma Runtimes could fetch them and configure the
+sidecars accordingly.
 
-```yaml
-runtimes:
-  - id: "1"
-    data: # arbitrary data
-      foo: bar
-  - id: "2"
-    data: # arbitrary data
-      bar: baz
-```
+## PoC scenario
 
-The application reloads configuration every 10 seconds. 
+The scenario will present two Kyma Runtimes using [Dapr HTTP Bindings](https://github.com/dapr/docs/blob/master/reference/specs/bindings/http.md) to reach two different Redis instances.
 
-## Details 
+1. Kyma Control Plane admin sets the configuration for two Kyma Runtimes.
+2. Runtime Agents from both Runtimes fetch the configuration and create the Dapr Bindings.
+3. Kyma Control Plane admin changes the configuration.
+4. Agents fetch the new configuration, apply it, and restart the Pods.
+5. Both Runtimes use the new configuration.
 
-### API
+As it's the Proof of Concept stage, we can use some temporary solutions, such as:
+- Configuration held in memory instead of a database
+- Pods being restarted by the Agent instead of a more complex solution
 
-The application exposes the following endpoints:
+## Reloading configuration of Dapr sidecars 
 
-- `GET /runtimes`
-   
-    Returns configuration for all Runtimes.  
-    
-    Possible error codes: `500`
-    
-    Example response:
-    ```json
-    [{"id": "1", "data": {"foo":  "bar"} }, {"id": "2", "data": {"bar":  "baz"}}]    
-    ```
+We need to configure Dapr sidecars in a Kyma Runtime from one central Kyma Control Plane. However, he Dapr sidecars have an [issue](https://github.com/dapr/dapr/issues/1172) concerning the fact that the whole Pod needs to restart for the configuration to be applied.
 
-- `GET /runtimes/{runtimeId}`
-   
-    Returns configuration for a given Runtime.  
-    
-    Possible error codes: `404`, `500`
-    
-    Example response:
-    ```json
-    {"id": "1", "data": {"foo":  "bar"} }    
-    ```
-  
-  
-## Development  
+There are a few solutions for this problem:
 
-### Build
+### Control Plane Runtime Agent will restart every Pod with a Dapr sidecar
+This solution may be the simplest, but it could add too many responsibilities for one component. A single Control Plane Runtime Agent would fetch the configuration, apply it, and then restart all Pods.
 
-To build the source code, navigate to the `./component` directory and run:
+Pros:
+- No new components in the Runtime 
+- Less resource-consuming
 
-```bash
-make build-image
-```
+Cons:
+- Too many responsibilites for one component
 
-### Run on a local machine
+### New component responsible for restarting Pods
+This solution adds a new component that will be triggered by the Agent and will restart Pods mentioned in the configuration.
 
-To run the application without building a Docker image, navigate to the `./component` directory and execute the following command:
+Pros:
+- Responsibilites distributed between components
 
-```bash
-go run cmd/main.go
-```
+Cons:
+- Another component will definitely mean more resource consumption, which we want to avoid
 
-The Runtime Governor listens on the `127.0.0.1:3001` port and reloads the configuration file from the `hack/config.yaml` path.
+### Dapr contribution that would allow configuration reloading
+This solution will surely be the most time-consuming. It seems like a really big change to the Dapr ecosystem and the approach to that is still being [discussed](https://github.com/dapr/dapr/issues/1172#issuecomment-610568718). We would have to create a proposal, wait for it to be accepted, and then develop the feature.
 
-## Installation
+Pros:
+- Configuration reload out-of-the-box
 
-To install Runtime Governor on a Kubernetes cluster using Helm 3 in the `kcp-poc` Namespace, run the following command:
-
-```bash
-LOCAL_ENV=false ./deploy.sh
-```
-
-The script will add a proper `runtime-governor.{DOMAIN}` entry to `/etc/hosts`.
-
-## Configuration
-
-You can use the following environment variables while running the `deploy.sh` script:
- - `LOCAL_ENV` - a variable that specifies whether the script runs on a local environment with Minikube (default value: `true`)
- - `DOMAIN` - a domain used on a cluster (default value: `kyma.local`)
- - `ISTIO_GATEWAY_NAME` - Istio Gateway name (default value: `compass-istio-gateway`)
- - `ISTIO_GATEWAY_NAMESPACE` - Istio Gateway Namespace (default value: `compass-system`)
-
-For example, to set a different domain and install Runtime Governor on a Kubernetes cluster, run the script in such a way:
-
-```bash
-LOCAL_ENV=false DOMAIN=foo.bar ./deploy.sh
-```
-
-## Testing
-
-To verify if Runtime Governor works properly, execute the following command, which returns all Runtime configurations from the mounted ConfigMap:
-
-```bash
-curl https://runtime-governor.${DOMAIN}/runtimes
-```
+Cons:
+- Tremendous amount of work needed
