+++ docs/internal/investigations/runtime-governor-poc/governor/component/Makefile
@@ -1,8 +1,8 @@
 APP_NAME = runtime-governor
-APP_PATH = docs/internal/investigations/runtime-governor-poc/component
+APP_PATH = docs/internal/investigations/runtime-governor-poc/governor/component
 ENTRYPOINT = cmd/main.go
 BUILDPACK = eu.gcr.io/kyma-project/test-infra/buildpack-golang-toolbox:v20200423-1d9d6590
-SCRIPTS_DIR = $(realpath $(shell pwd)/../../../../../)/scripts
+SCRIPTS_DIR = $(realpath $(shell pwd)/../../../../../../)/scripts
 
 include $(SCRIPTS_DIR)/generic_make_go.mk
