+++ docs/internal/investigations/runtime-governor-poc/governor/component/internal/handler/handler.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
-	"github.com/kyma-project/control-plane/docs/internal/investigations/runtime-governor-poc/component/pkg/apperror"
-	"github.com/kyma-project/control-plane/docs/internal/investigations/runtime-governor-poc/component/pkg/model"
+	"github.com/kyma-project/control-plane/docs/internal/investigations/runtime-governor-poc/governor/component/pkg/apperror"
+	"github.com/kyma-project/control-plane/docs/internal/investigations/runtime-governor-poc/governor/component/pkg/model"
 	"github.com/pkg/errors"
 )
